fix(build): approve Vite 8 bundle ceilings
Preserve each affected scene's prior headroom and keep the Sprite2D Y-sort implementation isolated under Rolldown.
