+++ packages/babylon-lite/src/sprite/sprite-2d.ts
@@ -14,7 +14,6 @@ import { resolveSpriteFrame } from "./shared/sprite-atlas.js";
 import type { Sprite2DCustomShader } from "./sprite-custom-shader.js";
 import { _getSpriteFxHook } from "./sprite-fx-hook.js";
 import type { Sprite2DYSortState } from "./sprite-2d-y-sort.js";
-import { _getSprite2DYSortHook } from "./sprite-2d-y-sort-hook.js";
 import type { SpriteBlendDescriptor } from "./sprite-blend.js";
 import { spriteBlendAlpha } from "./sprite-blend.js";
 
@@ -161,6 +160,28 @@ export interface Sprite2DLayer {
 export interface Sprite2DIndexHandleHooks {
     readonly removeIndex: (index: number, last: number) => void;
     readonly clear: () => void;
+}
+
+/** @internal Opaque hooks implemented by the opt-in Sprite2D Y-sort module. */
+export interface Sprite2DYSortHook {
+    readonly add: (layer: Sprite2DLayer, index: number) => void;
+    readonly remove: (layer: Sprite2DLayer, index: number, last: number) => void;
+    readonly clear: (layer: Sprite2DLayer, previousCount: number) => void;
+    readonly dirty: (layer: Sprite2DLayer, lo: number, hi: number) => void;
+    readonly upload: (device: GPUDevice, layer: Sprite2DLayer, instanceBuffer: GPUBuffer, uploadedVersion: number) => number | undefined;
+    readonly drawOrder: (layer: Sprite2DLayer) => Uint32Array | null;
+}
+
+let _sprite2DYSortHook: Sprite2DYSortHook | null = null;
+
+/** @internal Install the optional Sprite2D GPU-order hook. */
+export function _registerSprite2DYSortHook(hook: Sprite2DYSortHook): void {
+    _sprite2DYSortHook = hook;
+}
+
+/** @internal Return the optional Sprite2D GPU-order hook. */
+export function _getSprite2DYSortHook(): Sprite2DYSortHook | null {
+    return _sprite2DYSortHook;
 }
 
 /** Per-sprite init record passed to `addSprite2DIndex` / `updateSprite2DIndex`. */
