+++ packages/babylon-lite/src/sprite/sprite-pipeline.ts
@@ -3,10 +3,10 @@ import { U16 } from "../engine/typed-arrays.js";
 import { BU, SS, CW } from "../engine/gpu-flags.js";
 import type { EngineContext } from "../engine/engine.js";
 import type { Sprite2DLayer, SpriteBlendMode } from "./sprite-2d.js";
+import { _getSprite2DYSortHook } from "./sprite-2d.js";
 import type { SpriteLayerFx } from "./custom-shader-core.js";
 import { _getSpriteFxHook } from "./sprite-fx-hook.js";
 import { _getSpriteCoverageGammaHook } from "./sprite-coverage-gamma-hook.js";
-import { _getSprite2DYSortHook } from "./sprite-2d-y-sort-hook.js";
 import { DEPTH_INSTANCE_STRIDE_BYTES, PURE_2D_INSTANCE_STRIDE_BYTES } from "./sprite-2d.js";
 import { _getAlphaToCoverageResolver } from "../render/alpha-to-coverage-hook.js";
 import { wgsl } from "../shader/wgsl.js";
